Document sparse dictionary cardinality estimation
