+++ vortex-duckdb/src/exporter/mod.rs
@@ -7,6 +7,7 @@ mod cache;
 mod constant;
 mod decimal;
 mod dict;
+mod dict_cardinality;
 mod fixed_size_list;
 mod list;
 mod list_view;
